♻️Refactor: 전시 추천 리팩토링
* 🐛Fix: 전시 수정 오류 수정

* 🐛Fix: Blue-Green 네트워크 오류 수정

* :recycle:Refactor: 도즐 볼륨 제거

* :bug:Fix: 오류 수정

* :bug:Fix: 오류 수정

* :bug:Fix: 오류 수정

* :bug:Fix: Blue-Green 서버 오류 수정

* :bug:Fix: Blue-Green 서버 오류 수정

* :bug:Fix: Blue-Green 네트워크 오류 수정

* :bug:Fix: Blue-Green 컨테이너 오류 수정

* :recycle: Refactor: 작품 등록 로직 변경

* ✨Feat: 전시 참여자 관련 API 작업

* :recycle:Refactor: 포트 제거

* :card_file_box:Comment: 인증 오류 로그 수정

* :wrench:Settings: JWT 환경변수 추가

* :sparkles:Feat: 전시 참여자 관련 API 작업

* :recycle:Refactor: 코드 리뷰 기반 수정사항 반영

* :recycle:Refactor: 코드 포맷 적용

* :sparkles:Feat: 사용자 첫 로그인 여부 API 생성

* :recycle:Refactor: 소셜로그인 성공 Redirect URL 변경

* :bug:Fix: cookie 오류 수정

* :bug:Fix: cookie 오류 수정

* :bug:Fix: cookie 오류 수정

* :bug:Fix: cookie 오류 수정

* :bug:Fix: CD 오류 수정

* :recycle: Refactor: 디버깅 코드 추가

* :recycle:Refactor: 가장 인기 있는 사용자 조회 추가

* :bug:Fix: JPA 오류 수정

* :recycle:Refactor: 코드 포맷 적용

* :bug:Fix: 존재하는 작품 재등록 오류 수정

* :bug:Fix: 연령대 추천 오류 수정

* :bug:Fix: 연령대 추천 오류 수정

* :bug:Fix: 전시 수정 오류 수정

---------


* :bug:Fix: 전시 추천 오류 수정

* :recycle:Refactor: 지금 뜨는 크리에이터 코드 추가

* :recycle:Refactor: 전시 추천 리팩토링

* :recycle:Refactor: 사용자 첫 로그인 페이지 리다이렉트 수정

* :recycle:Refactor: 전시 조회 로그 수정

---------

+++ src/main/java/com/likelion13/artium/domain/user/service/UserServiceImpl.java
@@ -255,7 +255,9 @@ public class UserServiceImpl implements UserService {
         String oldImageUrl = user.getProfileImageUrl();
         String newImageUrl = s3Service.uploadFile(PathName.PROFILE_IMAGE, profileImage);
 
-        if (oldImageUrl != null && !oldImageUrl.isEmpty()) {
+        if (oldImageUrl != null
+            && !oldImageUrl.isEmpty()
+            && !oldImageUrl.startsWith("http://k.kakaocdn.net")) {
           s3Service.deleteFile(oldImageUrl);
         }
         user.updateProfileImageUrl(newImageUrl);
@@ -325,7 +327,9 @@ public class UserServiceImpl implements UserService {
     try {
       newImageUrl = s3Service.uploadFile(PathName.PROFILE_IMAGE, profileImage);
 
-      if (oldImageUrl != null && !oldImageUrl.isEmpty()) {
+      if (oldImageUrl != null
+          && !oldImageUrl.isEmpty()
+          && !oldImageUrl.startsWith("http://k.kakaocdn.net")) {
         s3Service.deleteFile(oldImageUrl);
       }
     } catch (Exception e) {
@@ -772,9 +776,9 @@ public class UserServiceImpl implements UserService {
           }
         };
 
-    add.accept("themes: " + themes, 0.60);
-    add.accept("moods: " + moods, 0.20);
-    add.accept("formats: " + formats, 0.20);
+    add.accept("themes: " + themes, 0.80);
+    add.accept("moods: " + moods, 0.10);
+    add.accept("formats: " + formats, 0.10);
 
     float[] acc = null;
     for (int i = 0; i < vecs.size(); i++) {
